chore: move fetchAppOpportunities into react-queries
